Add focus management to the mobile menu interactions
When opening and closing menu states the focus is hidden on the trigger
that moves out of view. This change addresses the issue by moving the
focus to the logical trigger of the panel that comes within view.
There's also a bit of cleanup to avoid setting/unsetting all transitions
on the sibling submenus and narrow the focus (pun intended) to only the
child of the parent trigger.